Use constructors from current context instead of last-loaded component
This adds a "contextId" property to the QMLContext class.  This is set
when the QMLContext is initially created for the root object of a new
Component that contains imports (e.g. a new Component file).

When "construct" is called, the current contextId is used to look up the
available constructors.